multi resolution example updated using https://github.com/alamboley/Citrus-Engine/commit/eb7e7f7d5a69e7facba845a7030bdfd77f1177a8
example of how to setup viewport modes.

add for example

_baseWidth = 480;
_baseHeigth = 320;
_viewportMode = ViewportMode.LETTERBOX;

to setup a letterbox version of the app.
_baseWidth/Height is required for LETTERBOX, NO_SCALE, FULLSCREEN but not for MANUALof course...
Manual requires manual definition of starling stage size and so on, which would be default anyway.

if _baseWidth is defined but not the other (or the other way round) we just set it to the stage equivalent ... its not likely to happen but it prevents errors I guess.

StarlingCitrusEngine has the scaleFactor getter setter, could be used to integrate the Utils.FindScaleFactor inside StarlingCitrusEngine instead of how it is currently for the demo.