advancedSounds
the sound would start at volume 1 and only on next update have the proper volumes - so a peak would be heard.
That is because on first frame the camera is not at its right place when the space is being updated.
To fix that, camera position should be updated BEFORE calling super.update(timeDelta) in the state's update.